backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
pycparser==2.22
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
